Refactor component organization and rename components.
