+++ components.json
@@ -13,7 +13,7 @@
   "aliases": {
     "components": "@/shared/components",
     "utils": "@/shared/lib/utils",
-    "ui": "@/shared/components/ui/",
+    "ui": "@/shared/components/ui/base/",
     "lib": "@/shared/lib",
     "hooks": "@/shared/hooks"
   },
